move data point types
