+++ packages/stable/src/api/dataPoints/dataPointsApi.ts
@@ -1,20 +1,18 @@
 // Copyright 2020 Cognite AS
 
 import { BaseResourceAPI } from '@cognite/sdk-core';
+import type { ItemsWrapper } from '@cognite/sdk-core';
 import type {
   DatapointAggregate,
   DatapointAggregates,
-  DatapointInfo,
   Datapoints,
   DatapointsDeleteRequest,
   DatapointsMonthlyGranularityMultiQuery,
   DatapointsMultiQuery,
   ExternalDatapointsQuery,
-  IgnoreUnknownIds,
-  ItemsWrapper,
   LatestDataBeforeRequest,
-  Timestamp,
-} from '../../types';
+} from './types';
+import type { DatapointInfo, IgnoreUnknownIds, Timestamp } from '../../types/common';
 
 export class DataPointsAPI extends BaseResourceAPI<
   DatapointAggregates | Datapoints
